raspi/serial-only/requirements.txt
pyserial==3.4
algebraic-data-types==0.1.1
